requirements.txt
python-dotenv
loguru
openai
fastapi
uvicorn
google-generativeai==0.8.3
APScheduler
PyYAML

